+++ sam/app/api/health.py
@@ -1,7 +1,7 @@
 """Health Check Module"""
 
 from fastapi import APIRouter
-from src import __version__
+from app import __version__
 
 router = APIRouter()
