+++ sam/app/main.py
@@ -8,9 +8,10 @@ from fastapi import FastAPI
 from sam2.build_sam import build_sam2
 from sam2.sam2_image_predictor import SAM2ImagePredictor
 
-from src import __version__
-from src.api import prediction, health
-from src.config import Settings, get_settings
+from app import __version__
+from app.api import prediction, health
+from app.config import Settings, get_settings
+from app.download_weights import download_weights
 
 
 def load_predictor(settings: Settings) -> SAM2ImagePredictor:
@@ -28,6 +29,8 @@ async def lifespan(local_app: FastAPI) -> AsyncGenerator[None, None]:
     yield
 
 prefix = get_settings().api_base_path
+
+download_weights()
 
 app = FastAPI(
     title = "Cytomine Segment Anything Server",
