+++ sam/app/utils/extract_img.py
